requirements.txt
asgiref
beautifulsoup4
blinker
certifi
charset-normalizer
click
Django
django-crispy-forms
django-phonenumber-field
django-unfold
Flask
greenlet
idna
importlib_metadata
itsdangerous
Jinja2
MarkupSafe
MouseInfo
phonenumbers
pillow
PyAutoGUI
PyGetWindow
PyMsgBox
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pyperclip
PyRect
PyScreeze
pytweening
pytz
pywhatkit
requests
rubicon-objc
soupsieve
sqlparse
typing_extensions
urllib3
urlopen
Werkzeug
whitenoise
wikipedia
zipp
